Move some BsonArray / BsonObject tests into BsonCodecTest
